inspector arrayofstring

+++ backend-app-landscape/src/main/java/com/example/applicationlandscape/Repository/LinkRepository.java
@@ -4,5 +4,5 @@ package com.example.applicationlandscape.Repository;
 import com.example.applicationlandscape.Entity.Link;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LinkRepository extends JpaRepository<Link, Long> {
+public interface LinkRepository extends JpaRepository<Link, Integer> {
 }
